+++ NEWS.md
@@ -1,4 +1,25 @@
-# teal.modules.general 0.4.0.9001
+# teal.modules.general 0.4.1.9020
+
+### Breaking changes
+
+- Removed the `table` object decoration in `tm_missing_data` and `tm_outliers` (#899).
+- Removed the `summary_table` and `test_table` object decoration in `tm_g_distribution` (#897).
+
+### Bug fixes
+
+- Fixed "Add to Report" functionality in `tm_outliers`, `tm_missing_data` and `tm_g_distribution` modules (#899 and #897).
+
+### Miscellaneous
+
+- Fixed deprecation warning for `tern::stack_grobs` in `tm_g_association` (#903).
+
+# teal.modules.general 0.4.1
+
+### Bug fixes
+* Fixes output is not updated when filter is added for the modules:  `tm_a_pca`, `tm_a_regression`, `tm_g_scatterplot`, `tm_g_association`, `tm_g_bivariate`, `tm_g_distribution`, `tm_g_response`, `tm_t_crosstable` (#870)
+
+### Enhancements
+* Added `remove_zero_columns` to `tm_t_crosstable` to add an option to remove empty columns (#890)
 
 # teal.modules.general 0.4.0